Share the chart geometry, headline feed, and demo chain id.
