+++ demos/extension/src/TradingPanel.tsx
@@ -1,5 +1,6 @@
 import type { PasskeyCredentialMetadata } from "@category-labs/mera";
 import { parseDecimalAmount } from "@category-labs/mera-demo-shared/amount";
+import { CHART_WINDOW_SECONDS } from "@category-labs/mera-demo-shared/chart";
 import {
   costBasisAfterBuy,
   costBasisAfterSell,
@@ -22,10 +23,7 @@ import {
   type EvmContext,
   fundAccount,
 } from "@category-labs/mera-demo-shared/network";
-import {
-  CHART_WINDOW_SECONDS,
-  PriceChart,
-} from "@category-labs/mera-demo-shared/PriceChart";
+import { PriceChart } from "@category-labs/mera-demo-shared/PriceChart";
 import {
   CASH_SYMBOL,
   formatCash,
