+++ demos/extension/src/config.ts
@@ -2,11 +2,9 @@ const RP_ID = "mera.category.xyz";
 const RP_ORIGIN = "https://mera.category.xyz";
 const PASSKEY_PAGE_URL = `${RP_ORIGIN}/demo/passkey.html`;
 const PASSKEY_CHANNEL = "mera-passkey";
-const DEMO_CHAIN_ID = 31337;
 const RECOVERY_VISIBLE_MS = 60_000;
 
 export {
-  DEMO_CHAIN_ID,
   PASSKEY_CHANNEL,
   PASSKEY_PAGE_URL,
   RECOVERY_VISIBLE_MS,
