+++ demos/shared/package.json
@@ -7,10 +7,12 @@
     "./AccountChip": "./src/AccountChip.tsx",
     "./amount": "./src/amount.ts",
     "./base.css": "./src/base.css",
+    "./chart": "./src/chart.ts",
     "./costBasis": "./src/costBasis.ts",
     "./hd": "./src/hd.ts",
     "./market": "./src/market.ts",
     "./network": "./src/network.ts",
+    "./news": "./src/news.ts",
     "./NewsTicker": "./src/NewsTicker.tsx",
     "./PriceChart": "./src/PriceChart.tsx",
     "./ui": "./src/ui.ts",
